Backend/api
Got most of the backend and api calls re done now I just need to cleanup the frontend.
