+++ server/controllers/toDoControler.js
@@ -1,67 +1,15 @@
 import ToDo from "../models/ToDoModel.js";
 import { v4 as uuidv4 } from "uuid";
-import { Op } from "sequelize";
 import { validateData } from "../utils/dataValidation.js";
-
-async function getFolders(req, res, next) {
-  try {
-    const userId = req.user.id;
-    //Find all realted folders
-    const folders = await ToDo.findAll({
-      where: {
-        folder: { [Op.ne]: "" },
-        userId: userId,
-      },
-      attributes: ["folder"],
-      group: ["folder"],
-    });
-    //Check if folders exist
-    if (!folders || folders.length === 0) {
-      return res
-        .status(404)
-        .json({ message: "No folders found", success: false });
-    }
-    res.status(200).json({ folders, success: true });
-  } catch (error) {
-    next(error);
-  }
-}
-
-async function deleteFolder(req, res, next) {
-  try {
-    const { FolderName } = req.body;
-    const userId = req.user.id;
-    const validate = [{ FolderName: FolderName }];
-    validateData(validate, [["FolderName", "string", 249]]);
-    //Delete folder
-    const folder = await ToDo.destroy({
-      where: {
-        folder: FolderName,
-        userId: userId,
-      },
-    });
-    //Check if folder exists
-    if (!folder || folder.length === 0) {
-      return res
-        .status(404)
-        .json({ message: "Folder not found", success: false });
-    }
-    res.status(200).json({ message: "Folder deleted", success: true });
-  } catch (error) {
-    next(error);
-  }
-}
 
 async function loadTasks(req, res, next) {
   try {
     const userId = req.user.id;
     const folderName = req.query.foldername;
     //Input validation
-    if (typeof folderName !== "string" || folderName.length > 249) {
-      return res
-        .status(400)
-        .json({ message: "Invalid format len or type", success: false });
-    }
+    console.log(folderName);
+    const validate = [{ folderName: folderName }];
+    validateData(validate, [["folderName", "string", 249]]);
     //Load tasks
     const tasks = await ToDo.findAll({
       where: {
@@ -71,7 +19,7 @@ async function loadTasks(req, res, next) {
       attributes: { exclude: ["UserId"] },
     });
     //Check if tasks exist
-    if (!tasks) {
+    if (!tasks || tasks.length === 0) {
       return res
         .status(404)
         .json({ message: "No tasks found", success: false });
@@ -87,11 +35,8 @@ async function deleteTask(req, res, next) {
     const { Task } = req.body;
     const userId = req.user.id;
     //Input validation
-    if (typeof Task !== "string" || Task.length !== 36) {
-      return res
-        .status(400)
-        .json({ message: "Invalid format len or type", success: false });
-    }
+    const validate = [{ Task: Task }];
+    validateData(validate, [["Task", "string", 36]]);
     //Delete task
     const task = await ToDo.destroy({
       where: {
@@ -100,7 +45,7 @@ async function deleteTask(req, res, next) {
       },
     });
     //Check if task exists
-    if (!task) {
+    if (!task || task.length === 0) {
       return res
         .status(404)
         .json({ message: "Task not found", success: false });
@@ -117,11 +62,8 @@ async function updateTaskComplete(req, res, next) {
     const userId = req.user.id;
 
     // Input validation
-    if (typeof task_id !== "string" || task_id.length !== 36) {
-      return res
-        .status(400)
-        .json({ message: "Invalid format len or type", success: false });
-    }
+    const validate = [{ task_id: task_id }];
+    validateData(validate, [["task_id", "string", 36]]);
 
     // Find the task to get its current Completed status
     const task = await ToDo.findOne({
@@ -130,14 +72,12 @@ async function updateTaskComplete(req, res, next) {
         userId: userId,
       },
     });
-
     // Check if task exists
-    if (!task) {
+    if (!task || task.length === 0) {
       return res
         .status(404)
         .json({ message: "Task not found", success: false });
     }
-
     // Toggle the Completed status
     const updatedTask = await ToDo.update(
       { Completed: !task.Completed },
@@ -148,9 +88,8 @@ async function updateTaskComplete(req, res, next) {
         },
       }
     );
-
     // Check if the update was successful
-    if (!updatedTask) {
+    if (!updatedTask || updatedTask.length === 0) {
       return res
         .status(500)
         .json({ message: "Failed to update task", success: false });
@@ -195,11 +134,4 @@ async function createTask(req, res, next) {
   }
 }
 
-export {
-  getFolders,
-  deleteFolder,
-  loadTasks,
-  deleteTask,
-  updateTaskComplete,
-  createTask,
-};
+export { loadTasks, deleteTask, updateTaskComplete, createTask };
