+++ server/routes.js
@@ -2,9 +2,8 @@ import { Router } from "express";
 
 import { login, register } from "./controllers/authControler.js";
 import { authenticate } from "./middleWare/authMiddle.js";
+import { getFolders, deleteFolder } from "./controllers/folderControler.js";
 import {
-  getFolders,
-  deleteFolder,
   createTask,
   loadTasks,
   updateTaskComplete,
@@ -20,9 +19,6 @@ router.get("/getFolders", authenticate, getFolders);
 router.get("/getTododata", authenticate, loadTasks);
 router.delete("/deleteFolder", authenticate, deleteFolder);
 router.delete("/deleteToDo", authenticate, deleteTask);
-router.get("/validate", authenticate, (req, res) => {
-  res.status(200).json({ isAuthenticated: true });
-});
 router.put("/updateToDo", authenticate, updateTaskComplete);
 
 export default router;
